feat(data): harden database schema - numeric prices, timestamptz, constraints, indexes (Phase 4)
- Change all financial columns from float to numeric(10,2) to avoid rounding errors
- Change all timestamp columns to timestamptz for timezone safety
- Add CHECK constraints on status columns (bread, buy_order, invoices)
- Add missing indexes on lookup columns (customer_id, status, uuid, etc.)
- Fix outbox query to use LIMIT + FOR UPDATE SKIP LOCKED for safe scaling
- Fix test status values to match lowercase CHECK constraint ('pending'/'processed')
- Fix docker-compose RabbitMQ: add RABBITMQ_ERLANG_COOKIE + bind mount for cookie
- Mark Phase 4 as complete in ARCHITECTURE_AUDIT.md

+++ broker/broker_test.go
@@ -23,7 +23,7 @@ func (r *brokerStubRepo) InsertBreadMaker(data.BreadMaker) (int, error)         
 func (r *brokerStubRepo) InsertBuyOrder(data.BuyOrder, []data.Bread) (int, error)     { return 1, nil }
 func (r *brokerStubRepo) InsertMakeOrder(data.MakeOrder, []data.Bread) (int, error)   { return 0, nil }
 func (r *brokerStubRepo) AdjustBreadQuantity(int, int) (bool, error)                  { return true, nil }
-func (r *brokerStubRepo) AdjustBreadPrice(int, float32) error                         { return nil }
+func (r *brokerStubRepo) AdjustBreadPrice(int, float64) error                         { return nil }
 func (r *brokerStubRepo) PasswordMatches(string, data.Customer) (bool, error)         { return true, nil }
 func (r *brokerStubRepo) GetAvailableBread() ([]data.Bread, error)                    { return nil, nil }
 func (r *brokerStubRepo) GetBreadByID(int) (data.Bread, error)                        { return data.Bread{}, nil }
@@ -32,7 +32,7 @@ func (r *brokerStubRepo) GetBuyOrderByID(int) (data.BuyOrder, error)            
 func (r *brokerStubRepo) GetBuyOrderByUUID(string) (data.BuyOrder, error)             { return data.BuyOrder{}, nil }
 func (r *brokerStubRepo) GetAllBuyOrders() ([]data.BuyOrder, error)                   { return nil, nil }
 func (r *brokerStubRepo) UpdateOrderStatus(string, string) error                      { return nil }
-func (r *brokerStubRepo) GetOrderTotalCost(int) (float32, error)                      { return 0, nil }
+func (r *brokerStubRepo) GetOrderTotalCost(int) (float64, error)                      { return 0, nil }
 func (r *brokerStubRepo) DeleteOutboxMessage(int) error                               { return nil }
 func (r *brokerStubRepo) InsertOutboxMessage(data.OutboxMessage) error                { return nil }
 func (r *brokerStubRepo) GetUnprocessedOutboxMessages() ([]data.OutboxMessage, error) { return nil, nil }
